+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.Kuit4AndroidPrivate"
         tools:targetApi="31">
         <activity
-            android:name=".FavoriteActivity"
+            android:name=".favorite.FavoriteActivity"
             android:exported="false" />
         <activity
             android:name=".detail.DetailActivity"
