+++ app/src/main/java/com/example/kuit4androidprivate/model/CardData.kt
@@ -1,0 +1,7 @@
+package com.example.kuit4androidprivate.model
+
+data class CardData(
+    val id: Int,
+    val imgUrl: String,
+    val imgInt: Int
+)
