--- app/src/main/java/com/example/kuit4androidprivate/model/HomeCardData.kt
@@ -1,6 +1,0 @@
-package com.example.kuit4androidprivate.model
-
-data class HomeCardData(
-    val id: Int,
-    val imgUrl: String
-)
