+++ app/src/main/res/layout/activity_detail.xml
@@ -2,7 +2,8 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:backgroundTint="#D9D9D9">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_detail_top"
@@ -31,8 +32,8 @@
             app:layout_constraintTop_toTopOf="@id/iv_detail_back" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <ImageView
-        android:id="@+id/iv_detail_home_food"
+    <androidx.viewpager2.widget.ViewPager2
+        android:id="@+id/vp_detail_card"
         android:layout_width="0dp"
         android:layout_height="149dp"
         android:layout_marginTop="88dp"
@@ -42,10 +43,35 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
+    <com.google.android.material.imageview.ShapeableImageView
+        android:id="@+id/siv_detail_index_box"
+        android:layout_width="44dp"
+        android:layout_height="19dp"
+        android:layout_marginStart="308dp"
+        android:layout_marginTop="134dp"
+        android:layout_marginEnd="5dp"
+        android:layout_marginBottom="7dp"
+        android:background="@color/black"
+        app:layout_constraintBottom_toBottomOf="@id/vp_detail_card"
+        app:layout_constraintEnd_toEndOf="@id/vp_detail_card"
+        app:shapeAppearance="@style/corner_rd_8" />
+
+    <TextView
+        android:id="@+id/tv_home_vp_index"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="1/3"
+        android:textColor="@color/white"
+        app:layout_constraintBottom_toBottomOf="@id/siv_detail_index_box"
+        app:layout_constraintEnd_toEndOf="@id/siv_detail_index_box"
+        app:layout_constraintStart_toStartOf="@id/siv_detail_index_box"
+        app:layout_constraintTop_toTopOf="@id/siv_detail_index_box" />
+
     <TextView
         android:id="@+id/tv_detail_store_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="25dp"
         android:fontFamily="@font/pretendardvariable"
         android:text="아워떡볶이"
         android:textColor="?attr/colorPrimary"
@@ -53,7 +79,7 @@
         android:textStyle="bold"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/iv_detail_home_food" />
+        app:layout_constraintTop_toBottomOf="@id/vp_detail_card" />
 
     <ImageView
         android:id="@+id/iv_detail_star"
@@ -89,6 +115,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="10dp"
+        android:backgroundTint="@color/white"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/iv_detail_star"
@@ -101,6 +128,7 @@
         android:id="@+id/vp_detail"
         android:layout_width="match_parent"
         android:layout_height="0dp"
+        android:layout_marginTop="8dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
